Unify the func_0805FA60 prototype

+++ include/mode_backupstat.h
@@ -153,7 +153,7 @@ void EnableBg(s32 a);
 void SetupBg(s32 bg, u8 charBase, u8 screenBase, u8 palette);
 u16 GetKeysRepeat(void);
 void ModeRequest(Mode* mode, s32 arg);
-void func_0805FA60(u8 a, void* b, s32 c, u8 d);
+void func_0805FA60(s32 a, void* b, s32 c, u8 d);
 void func_0805FA8C(u8 a, s16 b, s16 c);
 void func_0805FCB0(u8 a, u8 b, u32 c, const char* d);
 void func_08060598(void);
