+++ include/mode_dummy.h
@@ -54,7 +54,7 @@ void func_08006120(s32 a, u16 b);
 void func_08006184(s32 a, u16 b);
 u8 func_08006314(void);
 void UpdatePlayTime(void);
-void func_0805FA60(s32 a, void* b, s32 c, s32 d);
+void func_0805FA60(s32 a, void* b, s32 c, u8 d);
 void func_0805FA8C(s32 a, u16 b, u16 c);
 void func_0805FCB0(s32 a, s32 b, s32 c, const char* s);
 void func_08060598(void);
